Moved the MockEC2Computer class to its own file
